+++ src/lib/createBuilder.ts
@@ -4,7 +4,7 @@ import { EventHandler } from '../EventHandler'
 export const createBuilder = <R extends Element>(root: R) => ({
     events: (eventType: string) => ({
         select: (selector: string) => ({
-            listen: (listener: DelegationListener<any, any>, listenerOptions?: AddEventListenerOptions): EventHandler<R> => new EventHandler({
+            listen: (listener: DelegationListener<any, any, any>, listenerOptions?: AddEventListenerOptions): EventHandler<R> => new EventHandler({
                 root,
                 eventType,
                 selector,